Use group-by fields to scope backlog for Aggregation Events
This change extends the Event and EventDto classes by a mapping of group-by fields to values. The map is populated when processing an alert with an aggregation. We later refer to the map in order to select the correct set of messages for the notification backlog. I.e. the backlog contains only messages that match the key value(s) in the notification.

Fixes #6381