Fixing route leave prompt for focused mode on dashboards.
Prior to this change, the route leave prompt asking the user for
confirmation if the current view is dirty and the route is left was
unconditionally triggered for any route history action. This means it is
also triggered when query parameters are being added to the current URL.

With the introduction of the focused mode and its use during widget
editing, this lead to the phenomenon that the prompt was always
triggered when a widget on an unsaved dashboard was focused, e.g. for
editing.

This change is now adding a check to the `Prompt` component used to
generate the confirmation prompt, checking if the new location is a
different path than the current one. Therefore it is not triggered if
only the query parameters are changing.