Make new aggregation builder wizard the new default. (#10482)
This change is toggling the feature flag used for the new aggregation
builder wizard. It removes the `aggregation-wizard` feature flag and
introduces the `legacy-aggregation-wizard` feature flag. The new
aggregation builder wizard is now active by default. If the old
aggregation builder ui is desired, the `legacy-aggregation-wizard` can
be used like this to enable it:

```
FEATURES=legacy-aggregation-wizard yarn start
```