Separate integration and unit tests in frontend.
This PR is implementing some structural changes that separate runs for
unit and integration tests in the frontend. This allows us to:

  - prevent rejected promises from integration tests leaking to other
    tests, causing false negatives
  - tune timeouts separately, as integration tests tend towards
    requiring higher timeout due to the increased complexity

The expected end result is to make the frontend tests run more reliable
without having to disable the `CreateNewDashboard` integration test, so
we can continue improving that test and the category of integration
tests.