Replacing `views/Select` with `common/Select` in `FieldSortSelect`. (#10535)
This PR is replacing the select component for the `FieldSortSelect`. UI
wise there is not a big difference between these components and in the
future we want to rely on only one `Select` component. The common select
component can also be tested more easily with `react-select-event`,
which will help us with the `react-select` update.

With this PR we are also removing the option to clear select, since it did had no effect.