Add validation that a sort does not work on timebased charts
## Motivation
Prior to this change, the user could configure sort elements on the
aggregation widget. But it either had not effect or it only broke the
visualization on timebase graphic charts.

## Description
This change will introduce a validaton to check if the sort element was
introduced together with a non table visualization and a timebase row
groupby.