Implementa Cancellable Promise (#10525)
## Motivation
Prior to this change, we rely in our code on the cancellbale promise
interface from bluebird promise. But we want to use the native promise
implementation.

## Description
This change will add a wrapper class arround the promise from native
`window.fetch` which will set a cancel flag which can be accessed later
to check if a promise was cancelled.

## Note
This does not mean the then chain will be stopped. This cannot be done
with the native promise implementation. But we can check if someone
wants us to ignore the result.

Fixes #10488