Handle session expiration properly in frontend. (#10573)
* Handle session expiration properly in frontend.

Through the transition from superagent to the native fetch API (#10342),
a different behavior for request errors was introduced. A non-successful
request does not trigger a promise rejection automatically, so an
additional `catch` call in the promise chain is required in order to
handle e.g. session expirations (which result in a 401 for subsequent
requests) properly.

This change is correcting handling of errored responses, adds tests and
cleans up the `FetchError` class by removing explicit `console.error`
calls in its constructor, leaving this to the environment for uncaught
promise rejections, but adding proper `name` & `message` fields so the
class cleanly extends the `Error` type.

* Removing unused import.