Only disable widget query controls partially when dashboard filter is defined. (#10527)
* Do not throw error on dashboards with undeclared query parameters.

We've recently implemented a time range info for dashboard widgets. This
PR is fixing a runtime error which occured when no search types exist.
This is the case when a filter with an undeclared parameter   is being
applied.

* Only disable widget time range picker when dashboard time range filter is defined.

Previously we disabled all widget query controls when a dashboard
filter (time range or query) was defined.

* Display global query override next to widget query in widget edit mode,
to make clear that the global filter gets combined with the widget
query.

* Creating separate function to reset global override timerange.

* Changing reset filter button titles.

* Override info style adjustments

- Change query override input height
- Hide cursor for disabled query inputs
- adjust query override border radius

* Change width of query override info

* Updating tests

* Fixing import path

* Improving layout of `TimeRangeOverrideInfo`.

* Renaming `TimeRangeOverrideInfo` to `WidgetTimeRangeOverride`.

* Renaming `QueryOverrideInfo` to `WidgetQueryOverride`.

* Fixing linter warnings

* Fixing and improving `WidgetQueryControls.test`.