Centralize timeout multiplier handling.
In a couple of places we need to set custom timeouts for jest tests. In
these cases we should take the `TIMEOUT_MULTIPLIER` env variable into
account. In order to avoid having to handle that over and over again
(parsing, validating, applying it), this change is adding two functions
to the `jest-preset-graylog` module:

  - `timeoutMultiplier()`: returning the raw multiplier, `1.0` if not
    set or not parseable
  - `applyTimeoutMultiplier(x: number)`: applies the configured timeout
    multiplier to any given timeout

The latter function should be used in tests to apply the multiplier to
any custom timeout override.