During investigation of (#10584)
https://github.com/Graylog2/graylog2-server/issues/10168 it turned out
that, IF at least one of non-default outputs was requested to shut down,
THEN all available instances of `OutputBufferProcessor` get blocked (for
the period configured in `output_module_timeout`) while waiting for the
non-default outputs to hand over a message down the flow.

It causes troubles in the following discovered scenarios:

 1) an output is deleted from a stream
 2) output's configuration is changed, what implies output's restart

As result it leads to the default output (ES output) getting blocked
until the lock is released after reaching `output_module_timeout`. It
explains 'the visualization stops showing messages' mentioned in the
reported bug (no new messages in Elastic Search). Also if ring buffer's
capacity of OutputBuffer is fully exhausted, one can observe high CPU
utilization due to the blocked pipeline.

This commit simply counts down the latch where it was missed to do so.

TODO:

 1) Follow up with the reporter and sort out values of the following
 config props: `ring_size` and `output_module_timeout`
 2) Discuss whether it's a good idea to keep communication with the
 default and customer-provided outputs in the same thread

Signed-off-by: Danylo Iesaulov <danylo.iesaulov@graylog.com>