Allow to temporary empty relative range value inputs in date time picker.
Before this change it was not possible to temporary empty the relative
range value inputs. When a user tried to remove all numbers, we
automatically inserted a 1. We mainly did this because we are always
updating the form state `onChange`. If we would define for example
`undefined` for the relative range value, we would loose the
information which range type is selected. The range type is
being inferred based on the time range form value.

For example if the form state value for the relative range is 300, the
`RelativeRangeSelect` displays: 5 (value) minutes (type).

With this PR we are implementing a separate state for the `RelativeRangeSelect` which controls the
relative range inputs. We are still updating the form state
`onChange`. Every time the form state changes we are synchronizing the
`RelativeRangeSelect`. When a user empties the range value input, we are
defining `null` for the range form state. By implementing a separate state, we
are able to remember the previous selected range type and consider the
type when the user defines a completely new range value.