Fix issue rendering RawMessageLoader component
The `inputs` prop of `RawMessageLoader` now receives an `ImmutableJS.Map`
object but we didn't update usages of that prop. This resulted in an
incorrect behaviour when using the JS standard API with the
`ImmutableJS.Map` object, that could throw an error.

This change uses the ImmutableJS API to work with the `inputs` prop,
ensuring the code behaves as expected.

Fixes #10663