Make all WildcardPermission case sensitive
When introducing the new permissions system in 4.0
we started using object instead of string permissions.

We did not notice that WildcardPermissions are by default case insensitive.
For most cases that didn't make a difference,
but returning the self-editing permissions to a user with a
non-lowercase name would change from something like

`users:edit:HORST` to `users:edit:horst`

Which would break the frontend permission checks
for these cases.

-> Introduce a new CaseSensitiveWildcardPermission class
   that changes this default.