Reduce amount of clicks it needs to assign a role on user create.
Before this change the role select in the user create form was a multi
value select. Which means a user had to click once to select a relevant
role and once to submit the selection. With this change we are directly
submitting the selection when a user selects a role.

This solution is suitable for this case because we are not sending an
API request each time we are assigning a selected roles, like we do it on the user
edit page.