Fix widget config form validation on aggregation element remove. (#10697)
Before this change the widget config form validation was not being
executed when removing an aggregation element. This is only the case
when we define `validateOnChange` for the related `FormikArray` component.
We recently disabled `validateOnChange` for the `FormikArray`s (#10574), because a side effect
was that the validation is being executed twice, when changing a related
input. Because of this conditions the most pragmatic approach is using
`setFieldValue` instead of the `FomikArray` method `remove`. I also
tried defining a `validationSchema` for the form, but it has the same
side effects.