Fix date format in outgoing traffic graphs
## Motivation
Prior this change, we displayed the timestamps on the graph in the
timezone of the browser. That was because we gave plotly unix timestamps
without timezone information.

## Description
This change will use ISOString format to give plotly the desired
timezone (UTC). This leads to the effect that now time will be
displayed. Which is okay since it should describe a daily summary.

Fixes #10051