Define portal for aggregation element select dropdowns to fix overflow problem.
Before this change a dropdown of an element configuration select could
be cut off by their container. Opening a dropdown caused a scroll bar when
the select is at the bottom of the aggregation elements column and the
menu is too large.

With this change we are displaying the dropdown in a portal to display it
independently from its container.