Fix sorting of stream overview page (#10754)
## Motivation
Prior to this change, the backend send the streams in alphabetical order
but with regard of case sensitivity. The frontend would then sort the
streams again without case sensitivity which resulted in unsorted
looking streams.

## Description
This change will remove the second sorting on the frontend side and rely
on the backend sorting.