Improving overflow of select menus in aggregation elements column of aggregation wizard. (#10755)
* Always apply all widget changes when saving aggregation builder.

Before this change we were only saving changes, when clicking on the “Save” button in the aggregation builder, which already have been applied by clicking the:
- search submit button, for widget search controls changes (on a dashboard)
- “Apply changes” button, for widget aggregation element changes.

With this PR we are always saving every valid search controls and aggregation element change, when the user is saving the widget. Furthermore we:
- Renamed the aggregation elements “Apply Changes” button to “Update preview” and unified the background color with the search controls submit button.
- Renamed the widget edit “Save” button to “”Apply Changes”

This way it is easier to understand what effect each button has.

* Removing comment in `Widget.test`.

* Define portal for aggregation element select dropdowns to fix
overflow problem.

Before this change a dropdown of an element configuration select could
be cut off by their container. Opening a dropdown caused a scroll bar when
the select is at the bottom of the aggregation elements column and the
menu is too large.

With this change we are displaying the dropdown in a portal to display it
independently from its container.

* Define more dynamic menu placement for aggregation element selects.

* Updating usage of aggregation element selects in tests regarding select menu portals.