Use relative URLs for web interface assets served by Graylog server and backend API. (#10726)
* Use relative URLs for web interface assets served by Graylog server.

Prior to this change, all web interface assets were qualified according
from the following data points, in order of precedence:

 - The `X-Graylog-Server-URL` header of the request
 - The `http_external_uri` config setting
 - The `http_publish_uri` config setting

While this helps us supporting a couple of scenarios where the Graylog
server is accessed through a reverse proxy, it also produces issues when naming
schemes are different internally and externally and access from both
directions is desired. The only scenario which works in these cases is
configuring a valid `http_publish_uri`/`http_external_uri` for internal
and using the `X-Graylog-Server-URL` on the proxy for external clients.

This is related to all of these settings being absolute URLs, resulting
in generating absolute URLs for the web assets as well, unnecessarily.

In all of the cases this complexity of configuration would be
unnecessary, if the references to the web assets would be relative. This
would support all of the aforementioned use cases and does not rely on
proper configuration of any of the settings _in order to be able to load
the web assets properly_ if there is no additional path prefix.

Therefore, this PR is determining the effective external URI in the same
way as listed before, but takes only the path part of it to prefix all
web assets.

Fixes #10632
Fixes #10267

* Use relative URL for backend API as well.

* Removing unused check for trailing slash.

* Use relative URL for `index.html` in dev too.

* Use relative URL in documentation REST resources.

* Support relative URLs in shred library, used by swagger.

* Removing commented out section.