Fix race condition in lookup data adapter and cache lifecycle
Previously we have been using separate Server.Listener instances to
update the live state of adapters and caches as well as counting down a
latch to signal the startup code that all adapters and caches are ready.

Since the service listeners are not executed in any specific order, it
could happen that we count down the latches before the services have
actually finished starting.
In that case the lookup table couldn't bestarted because the adapters
and caches couldn't be found in the live state.

This change fixes the race condition by updating the live state and
counting down the latch in a synchronous way in a single listener
instance.

Fixes #10753