Fix closing dashboard tabs.
In #10739, the queries/query id structures were changed to ordered sets.
Unfortunately, this change was not reflected in the code that handles
closing dashboard tabs, so it called `indexOf` which is not present on an
ordered set.

This change is now fixing this by converting the set of indices to an
indexed sequence before calling `indexOf`. In addition, it disables the
menu item to close a dashboard tab if there is only a single dashboard
tab present.

Fixes #10834.