Do not hide legends for pie charts
## Motivation
Prior to this change, we wanted to hide the legend if only one label
was presented to the user. The evaluation of that state did not take
into account that the pie chat needs to always show the labels.

## Description
This change will add a neverHide flag to the PlotLegend component and
that flag will be used by the PieChart component.

Fixes #10837