Do not render column rollup label as value action in PlotLegend
## Motivation
Prior to this change, the Plot legend would render the column rollup
label as Value action and with that the user would get a menu offered
which would not really make sense.

## Description
This change will try to find out if the label it renders is a series
function and will not render it with Value Actions.

## Fixes also
This fixes also a problem that when multiple series where selected by
the user the value acitons did got presented even though they were
broken.

## Known Issue
If the series function will be renamed to a value which exists in the
chart as a value, the value is not clickable and the plot legend will
only display the label one time (as does plotly renders the it).

Fixes #10840