Decorator action menu cut off
## Motivation
Prior to this change, the decorator action menu was cut off to the left
side in the Message Table edit modus. The problem exists since we
display the edit modus in focus mode.

## Description
This change will remove the `pullRight` prop from the drop down
component to let the menu open up to the right side.

Fixes #10841