Always update all widgets after unfocusing
## Motivation
Prior to this change, we did only update all searches after we left the
focus mode from editing. But when we switched a message table to a log
view and unfocused the log view other widgets were not part of the newly
generated search result.

## Description
This change will update all widgets after leaving the unfocusing mode.