Generate new search when only `trend` setting is changed.
Prior to this PR, when enabling only the boolean `trend` setting of a numeric visualization and nothing else, we did not generate a new search and the `trend` search type was not retrieved. This lead to empty trending information. The root cause of this was that we generally leave out the `visualization` and `visualizationConfig` keys of an aggregation config when determining equality to decide if we need a new search or not.

This change is now implementing `equalsForSearch` methods in the `VisualizationConfig` classes (defaulting to return `true`, as visualization configs should generally not need to generate a new search) and in `NumberVisualizationConfig` (returning a boolean depending on the presence and value of the `trend` setting of both configs). This leads to generating a new search when the `trend` setting is flipped.

The issue can be easily reproduced:

  - start a new search/dashboard
  - create a `Message Count` predefined aggregation
  - edit it
  - toggle only(!) the `trend` setting and apply it
  - see that trending information is empty (`-- / --`)