Adding package index for `aggregationwizard` module.
This change adds a package index (`index.ts`) for the `aggregationwizard` module. The intention is that any other module outside of the `aggregationwizard` directory should only import exports from the package index. This allows us to create an explicit boundary between this module and the rest of the system, allowing us to refactor the module's internals while avoiding to break external dependencies.