Disable automatic rebasing for dependency update PRs. (#10876)
This change is disabling automatic rebasing for dependabot's update PRs.
The motivation for this is to reduce the number of running builds on our
CI due to rebased PRs, whenever one of the package management
configuration files are touched. Most of these rebases are unnecessary,
especially for PRs which take longer to review.

In order to trigger a rebase manually for PRs, a simple

```
@dependabot rebase
```

as a PR comment is enough.