Use proper Jest types.
This change is replacing the previously used Jest types (`@types/jest`)
with the official ones (`@jest/types`), supporting Jest 27.

Unfortunately, the `setSystemTime` function is not part of the type
signatures, so a `@ts-expect-error` directive is added to a) silence the
error, b) throw an error when this is fixed upstream.