Store messages that failed pipeline processing
This submits messages that experienced failures
during pipeline processing to the FailureHandlerService.

Introduce new `Indexable#getMessageId()` that returns the message ULID
instead of the deprecated `getId()` that returns a UUID.
While migrating everything away from the UUID to the ULID
is out of scope for this PR, we don't want to introduce new code using
UUIDs.