To not set 'to' when setting TimeRange to 'All Time'
## Motivation
Prior to this change, the backend raised an error if a user switch the
timerange from 'from -> to' to 'All time' since 'to' was still send to
the server, which the server refuesed to accept.

## Description
This change will change the check where the submitted timerange is
checked to be 'All Time'
Before we did not see the timerange as 'All Time' if 'to' was still set,
but to would have been unset AFTER it acknowleged the time range a s
'All Time'.