Fix returning count and grand_total of views
## Motivation
Prior to this change, when returning the list of saved searches
or dashboards the pagination info always had a count of zero
and the combined result of dashboards and saved searches count in
grand_total.

## Description
This PR will slightly change the pagination api so grand total also can
have a DBQuery which will be based on the type (search / dashboard) and
will initiate the newly created PaginatedList with the correct count.