Late acknowledgement of messages with processing failures
Before this change, messages with pipeline errors were acknowledged before
handing them over to the FailureHandlerService.

If the FailureHandlerService could not store the failures,
because for instance elastic search was down, we would lose
all messages that were in the submission queue, when we shut down
Graylog.

This can be avoided, by only acknowledging them once they are
successfully indexed.