Fix retrying indexing when connection failed to elastic search 7 (#11057)
## Motivation
Prior to this change, the creation of an bulk request retryer failed
with an exception if the connection to elasitc search 7 could not be
established. Since the Exection was wrapped in a new elastsic search
execption.

## Description
This change will check if the error which causes the retryer to be
executed is an IOException and  will not raise a new execption, but
instead retries in configured time.