Drop gl2_processing_error field from failed_message
The failed message already stores this error separately.
Having this field twice would be redundant and if we decide
to reprocess the failed message, it should also be empty.

This change also creates a copy of the failed message,
so the modification only happens for the failure indexer.

Copying the message also solves the problem that when the message
is furtherly altered by other processing steps that could follow
if  `keepFailedMessageDuplicate` is active.