Unify failure submission in MessageFilterChainProcessor
While being processed, a message could experience multiple
errors. One in processing and eventually another exception.
This would lead to two submitted failures for the same message.

Instead of submitting the exception failure directly,
we also add it to the gl2_processing_error field,
which will be picked up later in the ProcessBufferProcessor.