WIP: Store more detailed failure context on Messages
The goal of this PR is to have a dedicated failure_cause
which for processing errors, would always be an Exception class.

The message of the rootCause can be extracted when
the failure is submitted

Multiple erros on a single message will need to be
split up into multiple ProcessingFailure objects.