Use singleton instance of PipelineInterpreterState for IlluminateProcessor
Previously, the IlluminateProcessor was explicitly creating new instances of `PipelineInterpreter.State` for each processed message. This state instance is intended to be a singleton in the original design. This commit migrates the usage to a singleton reference that it reused for each processed message. It also maintains a reference to the current configuration, and should update as the configuration is updated.