Assign page titles dynamically instead of setting them during render.
Prior to this change, the tab/page titles were set during render in some cases. This results in a newly loaded existing dashboard being flagged as dirty immediately.

This change is now assigning page names based on the index in the queries list instead of setting the title explicitly.

Fixes #11156.