Shorten widget titles with ellipsis when overflowing available space.
Prior to this change, a widget's title was always shown in full, even
when there was not enough space, leading to unwanted line breaks and
even taking over space of the actual visualization.

This change is now implementing shortening of the widget's title and
adding an ellipsis when the available space is exceeded.