Fix warning about receive buffer size
We were previously warning when the receive buffer size reported
by netty wasn't exactly matching the receive buffer size that was
configured. However, as the OS can choose to overprovision the buffer,
we shouldn't warn if the reported buffer size is exceeding the expected
buffer size.

With this change, we are only warning when the reported buffer size is
smaller than what we were expecting, as this is the problematic case
worth warning the user about.

Fixes #7720