Checking if input node is local asynchronously in message details.
This change is a follow-up to #11207. This PR was fixing the symptoms of
the underlying issue, but not the reason. Prior to it, we were expecting
the nodes list to be present in the `NodesStore`, which does not
necessarily have to be the case, depending on timing. Unfortunately,
this check was performed synchronously, so there was no way to just wait
for the nodes list to arrive before returning a result.

This change is now making the `isLocalNode` check return a
`Promise<boolean>` instead, allowing us to wait for the nodes list in
case of its absence.