Fix race condition in message details related to uninitialized nodes store. (#11207)
This change fixes an issue that occurs when immediately expanding the
message details of an arbitrary message on the search page. This is
caused by the `MessageDetail` calling `isLocalNode`, which is currently
an alias for `NodesStore.getNode()`. If the nodes store was not
initialized (has not retrieved an initial nodes list) before this,
`NodesStore.getNode` throws an exception as it is accessing
`this.nodes[id]`.

Fixes #11089.