Checking if input node is local asynchronously in message details. (#11225)
* Fix race condition in message details related to uninitialized nodes store.

This change fixes an issue that occurs when immediately expanding the
message details of an arbitrary message on the search page. This is
caused by the `MessageDetail` calling `isLocalNode`, which is currently
an alias for `NodesStore.getNode()`. If the nodes store was not
initialized (has not retrieved an initial nodes list) before this,
`NodesStore.getNode` throws an exception as it is accessing
`this.nodes[id]`.

Fixes #11089.

* Checking if input node is local asynchronously in message details.

This change is a follow-up to #11207. This PR was fixing the symptoms of
the underlying issue, but not the reason. Prior to it, we were expecting
the nodes list to be present in the `NodesStore`, which does not
necessarily have to be the case, depending on timing. Unfortunately,
this check was performed synchronously, so there was no way to just wait
for the nodes list to arrive before returning a result.

This change is now making the `isLocalNode` check return a
`Promise<boolean>` instead, allowing us to wait for the nodes list in
case of its absence.