Passing positions to grid again, instead of through widgets. (#11291)
This change is effectively reverting a small change performed in #11244,
which changed our way of using `react-grid-layout` by passing widget
coordinates through widget props instead of a global positions object.

Unfortunately, the `getDerivedStateFromProps` method in the
`ReactGridLayout` component does not take the content of the `children`
props into account when determining if a new layout needs to be
calculated, just the children's keys. This leads to the fact that we
cannot update widget positions "from the outside" when passing
coordinates through children props.

Therefore, this change is using a global positions map again, making
sure that it's aligned with the current state of widgets by retrieving
it in one central point in the code.