Ensure filter events have "source_streams" set
When an event definition doesn't have any streams configured, this
previously resulted in an empty "source_streams" field in the event
which allowed all users to see the event.

To fix it, we are now using the same way to compute the "source_streams"
for filter events that we are using for aggregation events.

Fixes #11301