Ensure filter events have "source_streams" set (#11303)
When an event definition doesn't have any streams configured, this
previously resulted in an empty "source_streams" field in the event
which allowed all users to see the event.

To fix it, we are now using the same way to compute the "source_streams"
for filter events that we are using for aggregation events.

Fixes #11301

(cherry picked from commit cfe478eb01a09af38878de8a8fe36f90bb8f4264)