Handle empty responses on plaintext requests
Graylog server API returns JSON responses also when doing a plaintext
request. Sometimes though, it is possible that we get an empty response
from the server, which will throw an error when trying to parse the JSON
response using the JS `fetch` API.

Since the default response can read a JSON response from the server and
already handles the situation described above, this commit replaces the
custom response handler for plaintext responses with the default one.

Fixes #11343