Use query fields instead of all fields for aggregation wizard's field select (backport of #11405 fro 4.2) (#11414)
* Use query fields instead of all fields for aggregation wizard's field select (#11405)

* Use query fields instead of all fields for aggregation wizard's field select

Prior to this change, the options list used for the field select of the aggregation wizard was based on the "all" field types list. That one includes field types for all streams a user has access to.  Unfortunately, this list of streams is lacking e.g. the indexing failure stream. This can lead to problems like the one described in #11404.

This change is now generating the field options list from the field types of the currently selected query.

Fixes #11404.

* Providing query fields in tests.

* Fixing linter warnings.

Co-authored-by: Dennis Oelkers <dennis@graylog.com>