Remove not needed `UserTimezoneTimestamp` component and replace it with `Timestamp` component.
Since the `Timestamp` component displays the provided timestamp based on the user timezone by default, we do not need the `UserTimezone` component.
The only difference, when a user has no timezone configured, we are the `Timestamp` component will use the browser timezone as a fallback.
Replacing the `UserTimezoneTimestamp` with the `Timezone` component will unify the timestamp handling in some parts of the search.