Improve user experience when changing relative time ranges in date time picker. (#11302)
* Unify the way we display a readable relative time range in the date time
picker and search bar.

* Simplify handling of relative time ranges state in date time picker.

Before this change we maintained the ranges of a relative time range in the date time picker as number in seconds (e.g. 300).

To improve the user experience we classified the number and transformed it into a human
readable format (like 5 minutes) and implemented an input for the
value and a select for the unit. This approach had some disadvantages:

- classification happened in the relative range select input on every
  change.
- in some cases changing the value lead to a transformation of the range directly
  after the change, e.g. 120 minutes were transformed to 2 hours.
- a "from" range of 0 represented the "all time" range, which required a more
  complex state handling.
- a not defined "to" represented the "all time" range, which also
  required a more complex state handling.

With this change we are using a different format for relative time
ranges in (only) the date time picker: `{ type: "relative", from: { value: 5, unit: "minutes", isAllTime: false }, to: { value: 4, unit: "minutes" isAllTime: false } }`.
The benefits:

- we only need to classify the relative time range when opening the date
  time picker. When we submit the date time picker changes, we normalize
  the time range.
- the state handling inside the date time picker is simpler
- transforming the time range from e.g. 120 minutes to 2 hours only
   happens when submitting the date time picker changes.

* Add missing license

* Implement fragment to fix strange stylelint indentation error.

* Cleanup

* Always set "Until: now" when selecting "From: All Time".

* Simplify `classifyRange` function.

* Add missing default prop.