Add guard checks to event bus subscriptions in index field type poller (#11510)
* Add guard checks to event bus subscriptions in index field type poller

Without this a non-master node would start polling index field types for
newly created index sets. The periodical doesn't get started on
non-master nodes, but due to the implementation of the periodicals
system the class gets instantiated and thus subscribed to the event bus.

Fixes #11504

* Remove guard for handleIndexDeletion method

The "IndicesDeleteEvent" is a not a cluster event and should be executed
on non-leader nodes as well to ensure timely cleanup.