Include existing URLs in allow list (#11407)
* Include existing URLs in allow list

Before this change, changing the `newUrlEntry` prop in
`URLWhiteListFormModal` would reset the component state, removing any
previously configured URLs in the allow list.

This change ensures the component will keep the existing allow list when
being reset, and does also work for users with the creator role.

Fixes #11345

* Fix state issues with URLWhiteListFormModal

Replace class with functional component, avoiding some catches when
using `componentDidUpdate`. This also helps making clearer a few things
we did in the component, and what needs to be updated when the component
changes.

Refs #11345

* Add tests for URLWhiteListFormModal

* Add type to state

* Use same event options in `createEvent`

Currently the function uses two different ways to create events, depending
on the browser's support for the `Event` constructor. However, both ways
initialize events with different options, which results in different
behaviours, depending on how the event was created.

* Make `FormsUtils.triggerInput` more generic

Function now directly receives an HTML input, instead of an specific URL
whitelist input.

* Refactoring around input ref

- Type the ref
- Use provided `getInputDOMNode()` method to get to the right input

* Fix useEffect dependencies

* Fix linter