Refactor index field type polling (#11562)
* Refactor IndexFieldTypePollerPeriodical

Previously, the periodical was running at a low frequency and was
scheduling separate threads to poll indexes for field names at the
rate configured for the respective index set. These threads needed to
be maintained and would make it hard to stop running the periodical for
a while and then restart it, because the threads would need to be
stopped and re-scheduled accordingly.

The periodical will now run every second and poll those indexes for
field names which need a refresh, based on their configuration. The
periodical can be stopped by not executing it for a while. The downside
is, that the granularity of the index polling is reduced to a resolution
of seconds, but that should be OK.

* Don't re-initialize after index set creation/deletion

Instead of re-initializing everything, we just drop the list of index
sets so that upon the next periodical run, we will re-fetch a fresh
list and adjust the "last poll" state.

* Remove duplicate test which was added by accident

* Reorder methods

* Add configurable, periodic full refresh of index field types