Use surrounding timerange options directly in SurroundingSearchButton.
This means the order of keys in the search config will be used as-is, which is consistent with RangePresetDropdown (i.e. the relative timerange dropdown).  For most cases this will be in ascending order of duration length, but also opens up the possibility for users to use the API to specify a custom order.