Fail early if mongodb is unavailable
If the MongoDB isn't reachable during server startup,
the server will hang for about 60min before exiting with an error.

This is due to the fact, that Guice will not abort instantly when it
hits an error. For each dependency that needs a mongodb connection
a new attempt is made. The default connection timeout is 30s.

A way to improve this, is to test the mongodb connection
early during startup:

 - Extract MongoDB Binding into separate MongoDBModule
 - Add another hook into the CmdLineTool run() sequence
   that is called after the configuration is parsed, but before
   the full guice injector is created.
 - Request one MongoConnection in beforeInjectorCreation() so we
   can get an early exception at startup.